requirements.txt
langflow==1.2.0
langchain==0.3.10
openai==1.68.2
jupyter==1.1.1
python-dotenv==1.0.1
ipykernel==6.29.5